Create basic stack structure
Created a basic stack structure with service/pipeline separation. Also
started some shared props and some of the other context and env management
that we've done in Marble. Added linters and integrated Github actions to
run tests. Moved source to src and output to lib to make lint config easier.