Add Honeycomb service
This adds the infrastructure needed for the Honeycomb service. Summary of
resources added:
- Solr ECS services
- RabbitMq broker
- Honeycomb ECS rails service
- Honeycomb ECS sneakers service for processing image jobs in the
RabbitMq queues.
- CloudWatch Alarms to scale the sneakers service up/down based on jobs
in the image processing queues.

I've also made several changes to some of the patterns in the project:

**Dockerfiles under test**

In order to speed up testing I've changed all tests to point stacks at
fixtures for it's appDirectory. In these directories are dummy Dockerfiles
that pull from scratch. When unit testing our infrastructure code, we
really don't need to spend time building the actual Docker images. Ideally,
the creation of ContainerImages would be completely mocked out, but I had
issues getting this working, so the next best thing is to just point it
at the simplest Dockerfiles possible. Doing this meant I could also
remove the source actions from the Github actions.

**Missing assets in pipelines**

This is a multi-stack application whose pipelines will be exclusively
deploying components of the application. When we do this we can run into
issues related to the marshaling of application code that belongs to a
different pipeline. Ex: The buzz pipeline will not pull the app source
for honeycomb, but when the buzz pipeline calls cdk deploy, cdk will still
try to create the docker image for honeycomb and will fail to find it's
Dockerfile. To get around this I've added an AssetHelpers class. Instead
of constructing the DockerImageAsset and ContainerImage directly, now you
can just use this helper class to do it for you, ex:
```typescript
const task = new FargateTaskDefinition(this, 'TaskDefinition')
const asset = new DockerImageAsset(this, 'DockerImageAsset', {
  directory: props.appDirectory,
  file: 'docker/Dockerfile',
})
task.addContainer('Container', {
  image: ContainerImage.fromDockerImageAsset(asset),
  ...
}
```
you should now use the AssetHelper.containerFromDockerfile:
```typescript
const image = AssetHelpers.containerFromDockerfile(this, 'DockerImageAsset', {
  directory: props.appDirectory,
  file: 'docker/Dockerfile',
})
task.addContainer('RailsContainer', {
  image,
  ...
}
```
This will properly prevent you from deploying the stack if the Dockerfile
is not found, but will still allow cdk to synthesize the entire graph of
Constructs so that it can see the interdependent resources and properly
create the export/imports between stacks.

**Shared ALB etc**

I've moved the ALB, Cluster, and private Namespace to the foundation
stack in order to reuse the ALB across multiple components of this
application. For the moment, I've only written Honeycomb to use these
since we have a lot of moving parts in this repo at the moment. Will
look at refactoring the others to use at some point.