Refactor pipeline tests to DRY up CDKPipelineDeploy tests
The main purpose of this is to better separate the way we test pipelines
before we do more of this. We have a reusable CDKPipelineDeploy class
that is responsible for generating the PipelineProject used to perform
the cdk deploy commands. These can be a little messy to test since it's
dealing with stringified jsons for buildspecs, so testing these can be
brittle and error prone. Ideally we should test this class to make sure
that cdk command is properly generated given a set of inputs, then pipeline
tests should only need to check that it calls the CDKPipelineDeploy with
the correct properties for its test and production deploy actions. This
should eliminate repeatedly writing pipeline tests that are really just
testing the behavior of this class.

Changes:
- Add placeholders for testing CDKPipelineDeploy class.
- Refactored the buzz pipeline tests as an example of mocking out and
testing the usage of the CDKPipelineDeploy class in the buzz pipeline.
- Mocking out a locally included class is a bit tricky in js. Added the
givens module to do lazy evals like in rspec. This allows us to dry up
some of the Construct instantiations similar to the way we were using the
stack() function, while still allowing us to instantiate a new subject
after mocking is done within the test scope.
- Changed the setup in the tests to have more explicit and unique values
to allow for more accurately testing how those passed in props are used.
- While refactoring the tests, found several inconsistencies in usage of
env in the buzz pipeline that was causing some bugs in the pipeline with
things like createDns, networkStackName, domainStackName. There was a lot
of overlap between the env object and some of the props being required.
- Also found an error with the pipeline not using the hostname prefix from
props. This would have caused collisions in dev with the dns record since
it was previously a constant of buzz-test and buzz.
- Removed mock credentials helper. Doesn't appear to be needed.