Fix specs
- FoundationStack now expects a honeycomb hostname. Added to all tests
that instantiate this stack
- BuzzPipeline now uses a TARGET_HOST env variable for its command in
SmokeTests so that we can reuse the project in the Test and Prod stages.
Modified tests to check this wiring
- Several stacks have changed to put their logs within the log bucket
using the entire domain name as the prefix in the bucket (which includes
an import from the foundation stack). Fixed tests to match this
- Beehive stack tests were reading from context for properties. Changed
to use props declared in the tests to better insulate them from changes
to code/configuration.
- Buzz pipeline now has smoke tests after prod. Added to tests.