Add Honeycomb Pipeline
This is primarily to add a deployment pipeline for Honeycomb, but required
making some changes to other thing for consistency. Changes:
- Changed buzz pipeline to use the namespace when mapping ssm parameters
to better adhere to separation of stacks in dev
- Moved handling of permissions for the migration project into the migration
project since it should be able to handle this itself
- There were some differences between buzz and honeycomb in the env keys
and ssm paths for reading database properties. Made some changes to both
to be more consistent
- Updated Buzz to use the latest Fargate platform. Likely doesn't need it
but did so for consistency and to make for less maintenance in the future
when this becomes the default
- Expanded migration project to allow passing additional env and commands
that need to run before migration. Needed this in order to recreate some
of the configuration that happens inside of the Honeycomb container when
it's built.
- Did a bit of refactoring of the namespaced policies to pare down some
policies, and make others a bit easier to maintain