Adding rate limit and scale to Honeycomb
Add rate limiting for dynamic content requests, such as API requests, in
order to protect the health of the application. Also adding a scale up
at 90% of those limits. We may find we need to scale up sooner than this,
but it's unclear what to expect for Honeycomb at the moment.