Create reusable Rails pipeline
The current implementation of our Buzz and Honeycomb pipelines rely on
the base rails container image when running migration commands. This
won't generally work since a migration may also include loading app
specific models which will not be in a generic rails image. In addition,
there may be additional commands that are run inside of the application
container's Dockerfile that would have to be recreated when running the
migration, leading to potential differences between the migration env
and the service env. This changes those pipelines to first build all
container images required for the service, then use the built rails
image in both the migration action and in the ECS task that is deployed.

A note about image asset management

Cdk wants to manage the container assets itself by pushing all containers
to a `aws-cdk/assets` ECR. It does not provide an easy way to instead push
to a different ECR, or use a different tag, things that are necessary for
referencing the correct image in the deploy and migration steps. [Cdk pipelines](https://aws.amazon.com/blogs/developer/cdk-pipelines-continuous-delivery-for-aws-cdk-applications/)
do a similar separation of build vs publish, but the manifest it generates
does not provide an easy way to identify which image is a rails image vs
an nginx image. So, I had to make changes the stacks that are deployed
to allow for specifying an override. To make this easier, I've added a
`AssetHelpers.getContainerImage` method that allows specifying a set of
context keys to check for where, if they are given, it will pull from
the given ECR/tag instead of building from a Dockerfile. For example,
the honeycomb stack will now look for `-c honeycomb:RailsEcrName=...`
and `-c honeycomb:RailsEcrTag=...` when deploying and, if given, will
use this for the ECS task's image. The pipeline will then add this to
the context overrides when it performs the `cdk deploy` to point it at
the Rails image it built in the build step. I would have liked to find
a more automatic way to handle this between the stack and pipeline, but
couldn't really think of a way that would not introduce the stack as a
direct dependency of the pipeline.

Additional improvements/changes

- The build stage will perform a check against the ECR to see if a tag
already exists for the sha that triggered the pipeline and will skip
building/pushing to ECR if it already exists. This also allowed us to
configure ECR to use immutable tags. This will help with tracing what
image is actually deployed to ECS, something that would not be possible
if we just moved tags.
- We had some duplication of efforts in where we pulled the Dockerhub
secrets. This changes all of these calls to use the secrets from
Secrets Manager.
- We commonly create a source artifact and action with the same ids.
Created a simple construct to wrap up this behavior for us

Potential for reuse

I've put a lot of the constructs in their own pipeline-constructs directory
to further separate them from the rest of the project, with the intent
to pull these out into something like ndlib-cdk in the future. At this
point, the only remaining dependencies these have on dec-blueprints
specific constructs are in the includes with rails-pipeline:
```typescript
import { CustomEnvironment } from '../custom-environment'
import { DockerhubImage } from '../dockerhub-image'
```
So these will also need to get lifted, or we'll have to figure out a
better separation for the values it needs from these.