Fix host configuration
There was a problem when deploying via pipelines that was preventing test
services from talking to other test services. Ex: honeycomb-test was
pushing images to honeypot-prod instead of honeypot-test. Changed honeycomb
to add context overrides for the hostname prefixes. It will get these values
from the other pipelines, since that information is really only known by
the other pipelines. It may make sense at some point to make a stage type
that, when given to a stack, allows the service stack to determine the
full hostname. We could then do away with requiring passing this info
around via pipelines, but that's a slightly bigger change then I want to
look at for this.