Fix beehive hostname and bump rails resources
Fixed an incorrect hostname for beehive. Bump resources on the rake task
by default to make reindexing the site faster (this can still be
overridden when creating a task). Bumped the resources for the rails
service. The site seemed a bit more sluggish than the live site for
the larger collections. This puts it a bit closer to the current performance.