Configure service endpoints
This is primarily to change pipelines to point all services at the correct
endpoints. Beehive test needs to point to the correct Honeycomb test
service to get collection data. Honeycomb needs to point to the correct
Beehive, Honeypot, and Buzz service for the same stage. Etc. These changes
ensures that the pipelines pass the correct endpoints into env at build
time so that these can be configured correctly. Previously, I was handling
this by sharing this information between pipelines, but Honeycomb and
Beehive now have a cyclic dependency, so I've moved the logic for creating
hostnames a bit earlier in the call stack by creating a PipelineHostnames
data structure that we can just pass this in to each stack that needs it.

In addition, there are some bug fixes:
- Beehive now uses an S3 Bucket deployment, but was not handling the case
when the assets weren't found such as in the other pipelines. Added
an AssetHelper to handle S3 deployments. When missing, it will add an
error to the stack, but will allow synth to continue by zipping up an
empty directory. It depends on /var/empty, which feels a bit heavy handed
and is not OS agnostic, but not entirely sure of a better way to resolve
this one at the moment.
- Found a bug with creation of hosted zones. It would have continued to
create a hosted zone even in libnd. Changed to prevent this and changed
code that depended on hosted zone permissions to first check if one was
created before adding a permission.
- Fixed a bug within context that was causing beehive to watch a different
infrastructure branch from other pipelines.
- Fixed an inconsistency in the beehive approval message. Updated it to
be more in line with our current approval message format.
- Added additional comments to some properties here and there.